Gui (#3)
* first draft

* first draft

* first draft state machine. Not working

* first draft state machine. Not working

* GUI working without functionnality

We can get in mode for QR or 2d figur, but we don't analyse anythin.
It's now a static picture. Should be a feed video.

* GUI working without functionnality

We can get in mode for QR or 2d figur, but we don't analyse anythin.
It's now a static picture. Should be a feed video.

* state have mode

* state have mode

* add pictures for 2d figur

* change size button for new image size

* switching between stream and hd picture work but need remove sleep

* adding the distance analyse. still need to warp image

* 2d figur and qr work but not tested with light

* camera control is not taking new hd picture every time

* adjusting gui and servo controller